Rename ocm-api-mock to generic api-mock
