+++ Makefile
@@ -47,7 +47,7 @@ WEBHOOK_PORT?=8080
 IMAGE_ORG?=quay.io/app-sre
 ADDON_OPERATOR_MANAGER_IMAGE?=$(IMAGE_ORG)/addon-operator-manager:$(VERSION)
 ADDON_OPERATOR_WEBHOOK_IMAGE?=$(IMAGE_ORG)/addon-operator-webhook:$(VERSION)
-OCM_API_MOCK_IMAGE?=$(IMAGE_ORG)/ocm-api-mock:$(VERSION)
+API_MOCK_IMAGE?=$(IMAGE_ORG)/api-mock:$(VERSION)
 
 # COLORS
 GREEN  := $(shell tput -Txterm setaf 2)
@@ -247,7 +247,8 @@ test-unit: generate
 
 ## Runs the Integration testsuite against the current $KUBECONFIG cluster
 test-integration: config/deploy/deployment.yaml \
-	config/deploy/webhook/deployment.yaml
+	config/deploy/webhook/deployment.yaml \
+	config/deploy/api-mock/deployment.yaml
 	@echo "running integration tests..."
 	@go test -v -count=1 -timeout=20m ./integration/...
 .PHONY: test-integration
@@ -300,7 +301,7 @@ dev-setup: | \
 	create-kind-cluster \
 	setup-olm \
 	setup-okd-console \
-	setup-ocm-api-mock
+	setup-api-mock
 .PHONY: dev-setup
 
 ## Setup a local env for integration test development. (Kind, OLM, OKD Console, Addon Operator). Use with test-integration-short.
@@ -385,12 +386,12 @@ load-addon-operator-webhook: build-image-addon-operator-webhook
 .PHONY: load-addon-operator-webhook
 
 ## Load OCM API mock images into kind
-load-ocm-api-mock: build-image-ocm-api-mock
+load-api-mock: build-image-api-mock
 	@source hack/determine-container-runtime.sh; \
 		$$KIND_COMMAND load image-archive \
-			.cache/image/ocm-api-mock.tar \
+			.cache/image/api-mock.tar \
 			--name=$(KIND_CLUSTER_NAME);
-.PHONY: load-ocm-api-mock
+.PHONY: load-api-mock
 
 # Template deployment for Addon Operator
 config/deploy/deployment.yaml: FORCE $(YQ)
@@ -398,9 +399,9 @@ config/deploy/deployment.yaml: FORCE $(YQ)
 		config/deploy/deployment.yaml.tpl > config/deploy/deployment.yaml
 
 # Template deployment for OCM API Mock
-config/deploy/ocm-api-mock/deployment.yaml: FORCE $(YQ)
-	@yq eval '.spec.template.spec.containers[0].image = "$(OCM_API_MOCK_IMAGE)"' \
-		config/deploy/ocm-api-mock/deployment.yaml.tpl > config/deploy/ocm-api-mock/deployment.yaml
+config/deploy/api-mock/deployment.yaml: FORCE $(YQ)
+	@yq eval '.spec.template.spec.containers[0].image = "$(API_MOCK_IMAGE)"' \
+		config/deploy/api-mock/deployment.yaml.tpl > config/deploy/api-mock/deployment.yaml
 
 # Template deployment for Addon Operator Webhook
 config/deploy/webhook/deployment.yaml: FORCE $(YQ)
@@ -426,16 +427,16 @@ endif
 .PHONY: setup-addon-operator
 
 ## Loads and installs the OCM API Mock into the currently selected cluster.
-setup-ocm-api-mock: load-ocm-api-mock \
-	config/deploy/ocm-api-mock/deployment.yaml
-	@echo "installing ocm-api-mock $(VERSION)..."
+setup-api-mock: load-api-mock \
+	config/deploy/api-mock/deployment.yaml
+	@echo "installing api-mock $(VERSION)..."
 	@(source hack/determine-container-runtime.sh; \
-		kubectl apply -f config/deploy/ocm-api-mock; \
-		echo -e "\nwaiting for deployment/ocm-api-mock..."; \
-		kubectl wait --for=condition=available deployment/ocm-api-mock -n ocm-api-mock --timeout=240s; \
+		kubectl apply -f config/deploy/api-mock; \
+		echo -e "\nwaiting for deployment/api-mock..."; \
+		kubectl wait --for=condition=available deployment/api-mock -n api-mock --timeout=240s; \
 		echo; \
 	) 2>&1 | sed 's/^/  /'
-.PHONY: setup-ocm-api-mock
+.PHONY: setup-api-mock
 
 ## Loads and installs the Addon Operator Webhook into the currently selected cluster.
 setup-addon-operator-webhook: $(YQ) load-addon-operator-webhook \
