+++ config/deploy/api-mock/00-namespace.yaml
@@ -1,4 +1,4 @@
 apiVersion: v1
 kind: Namespace
 metadata:
-  name: ocm-api-mock
+  name: api-mock
